[Behat] [DX] Moving attributes page to keep consistency
