+++ src/Sylius/Behat/Context/Ui/Admin/ManagingProductAttributesContext.php
@@ -13,8 +13,8 @@ namespace Sylius\Behat\Context\Ui\Admin;
 
 use Behat\Behat\Context\Context;
 use Sylius\Behat\Page\Admin\Crud\IndexPageInterface;
-use Sylius\Behat\Page\Admin\Product\Attribute\CreatePageInterface;
-use Sylius\Behat\Page\Admin\Product\Attribute\UpdatePageInterface;
+use Sylius\Behat\Page\Admin\ProductAttribute\CreatePageInterface;
+use Sylius\Behat\Page\Admin\ProductAttribute\UpdatePageInterface;
 use Sylius\Behat\Service\CurrentPageResolverInterface;
 use Sylius\Component\Product\Model\AttributeInterface;
 use Webmozart\Assert\Assert;
